+++ application/views/oneitemeditable.php
@@ -1,0 +1,6 @@
+<tr>
+    <!-- MODIFY the next line -->
+    <td>{Category}</td>
+    <td><a href="/customization/edit/{id}"><input type="button" value="{EquipmentID}"/></a></td>
+    <td>{Name}</td>
+</tr>