+++ application/views/oneitemnoteditable.php
@@ -1,0 +1,5 @@
+<tr>
+    <td>{Category}</td>
+    <td>{EquipmentID}</td>
+    <td>{Name}</td>
+</tr>